Likes Added
- Ability to like/unlike posts now added
- Likes past 1000 will be abbreviated
- Minor bug fixes and improvements

+++ iOS/Mumble/Mumble/Mumble.h
@@ -14,8 +14,8 @@
 @property (strong, nonatomic) NSString *content;
 @property (strong, nonatomic) NSString *createdAt;
 @property (strong, nonatomic) NSString *msgLocation;
-@property int comments;
-@property int likes;
+@property long comments;
+@property long likes;
 
 @property (nonatomic) double cellHeight;
